Move apphosting tests (#7318)
