+++ src/apphosting/secrets/index.spec.ts
@@ -1,14 +1,14 @@
 import { expect } from "chai";
 import * as sinon from "sinon";
 
-import * as apphosting from "../../../gcp/apphosting";
-import * as secrets from "../../../apphosting/secrets";
-import * as iam from "../../../gcp/iam";
-import * as gcb from "../../../gcp/cloudbuild";
-import * as gce from "../../../gcp/computeEngine";
-import * as gcsmImport from "../../../gcp/secretManager";
-import * as utilsImport from "../../../utils";
-import * as promptImport from "../../../prompt";
+import * as apphosting from "../../gcp/apphosting";
+import * as secrets from ".";
+import * as iam from "../../gcp/iam";
+import * as gcb from "../../gcp/cloudbuild";
+import * as gce from "../../gcp/computeEngine";
+import * as gcsmImport from "../../gcp/secretManager";
+import * as utilsImport from "../../utils";
+import * as promptImport from "../../prompt";
 
 describe("secrets", () => {
   let gcsm: sinon.SinonStubbedInstance<typeof gcsmImport>;
